Disable logs and sessionId
Unfortunately not working well enough on deployment to justify pushing to release. Also likely to end up with 1000s of files on the live version if we do not set up an automatic clean-up rule. We will re-integrate in collaboration with the new deployment process being set up, with logging directories/day, deleted weekly.